+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - Adds interaction tests to `MultiSelect` component.
+- Increases the z-index of the `FeedbackBox`'s `Drawer` component so it displays above all other elements on a page.
 
 ## 3.3.2 (September 19, 2024)
 
@@ -18,7 +19,6 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Lowercases the 'r' of '(required)' in labels and placeholders where necessary in order to align with sentence case guidelines.
 - Updates the `CheckboxGroup`, `DatePicker`, `FeedbackBox`, `Label`, `RadioGroup`, `SearchBar`, `Select`, `Slider`, and `TextInput` component to use a lowercase 'r' for '(required)' in labels and placeholders in order to align with sentence case typography guidelines.
-
 
 ### Fixes